pyproject.toml
[project]
name = "macinventory"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
]
